Fix test descriptions and move them into the correct folder
